Dmitry commit with Search function
