+++ src/App.jsx
@@ -1,19 +1,22 @@
+import { useState } from 'react';
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Recipes from "./components/Recipes.jsx";
-import Header from "./components/header.jsx"; 
+import Header from "./components/header.jsx";
 import Hero from "./components/hero.jsx";
-import Gradesite from "./components/gradesite.jsx"; 
-import Categories from "./components/categories.jsx"; 
+import Gradesite from "./components/gradesite.jsx";
+import Categories from "./components/categories.jsx";
 
 function App() {
+  const [searchQuery, setSearchQuery] = useState('');
+
   return (
     <Router>
       <div className="App">
-        <Header />
+        <Header onSearch={setSearchQuery} />
         <Routes>
-          <Route path="/" element={<><Hero /><Recipes /></>} /> 
+          <Route path="/" element={<><Hero /><Recipes searchQuery={searchQuery} /></>} />
           <Route path="/gradesite" element={<Gradesite />} /> 
-          <Route path="/categories" element={<Categories />} /> 
+          <Route path="/categories" element={<Categories />} />
         </Routes>
       </div>
     </Router>
