+++ src/components/Header.jsx
@@ -1,0 +1,147 @@
+'use client'
+
+import { useState } from 'react';
+import {
+    Dialog,
+    DialogPanel,
+    Disclosure,
+    DisclosureButton,
+    DisclosurePanel,
+    Popover,
+    PopoverButton,
+    PopoverGroup,
+    PopoverPanel,
+} from '@headlessui/react';
+import {
+    ArrowPathIcon,
+    Bars3Icon,
+    ChartPieIcon,
+    CursorArrowRaysIcon,
+    FingerPrintIcon,
+    SquaresPlusIcon,
+    XMarkIcon,
+} from '@heroicons/react/24/outline';
+import { ChevronDownIcon, PhoneIcon, PlayCircleIcon } from '@heroicons/react/20/solid';
+import { Link } from "react-router-dom";
+
+const products = [
+    { name: 'Analytics', description: 'Get a better understanding of your traffic', href: '#', icon: ChartPieIcon },
+    { name: 'Engagement', description: 'Speak directly to your customers', href: '#', icon: CursorArrowRaysIcon },
+    { name: 'Security', description: 'Your customers’ data will be safe and secure', href: '#', icon: FingerPrintIcon },
+    { name: 'Integrations', description: 'Connect with third-party tools', href: '#', icon: SquaresPlusIcon },
+    { name: 'Automations', description: 'Build strategic funnels that will convert', href: '#', icon: ArrowPathIcon },
+];
+const callsToAction = [
+    { name: 'Watch demo', href: '#', icon: PlayCircleIcon },
+    { name: 'Contact sales', href: '#', icon: PhoneIcon },
+];
+
+export default function Header({ onSearch }) {
+    const [searchQuery, setSearchQuery] = useState('');
+    const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
+
+    const handleSearchChange = (e) => {
+        setSearchQuery(e.target.value);
+        onSearch(e.target.value);
+    };
+
+    return (
+        <header className="bg-light-gray">
+            <nav aria-label="Global" className="mx-auto flex max-w-7xl items-center justify-between p-6 lg:px-8">
+                <div className="flex lg:flex-1">
+                    <Link to="/" className="-m-1.5 p-1.5">
+                        <span className="sr-only">Your Company</span>
+                        <img alt="" src="https://tailwindui.com/img/logos/mark.svg?color=indigo&shade=600" className="h-8 w-auto" />
+                    </Link>
+                </div>
+                <div className="flex lg:flex-1 justify-center">
+                    <input
+                        type="text"
+                        placeholder="Search recipe name or ingredient"
+                        value={searchQuery}
+                        onChange={handleSearchChange}
+                        className="border rounded-lg p-2 w-1/2"
+                    />
+                </div>
+                <div className="flex lg:hidden">
+                    <button
+                        type="button"
+                        onClick={() => setMobileMenuOpen(true)}
+                        className="-m-2.5 inline-flex items-center justify-center rounded-md p-2.5 text-gray-700"
+                    >
+                        <span className="sr-only">Open main menu</span>
+                        <Bars3Icon aria-hidden="true" className="h-6 w-6" />
+                    </button>
+                </div>
+                <PopoverGroup className="hidden lg:flex lg:gap-x-12">
+                    <Link to="/" className="text-sm font-semibold leading-6 text-gray-900">
+                        Startsida
+                    </Link>
+                    <Link to="/categories" className="text-sm font-semibold leading-6 text-gray-900">
+                        Kategorier
+                    </Link>
+                    <Link to="/gradesite" className="text-sm font-semibold leading-6 text-gray-900">
+                        Betygsättning
+                    </Link>
+                </PopoverGroup>
+            </nav>
+            <Dialog open={mobileMenuOpen} onClose={setMobileMenuOpen} className="lg:hidden">
+                <div className="fixed inset-0 z-10" />
+                <DialogPanel className="fixed inset-y-0 right-0 z-10 w-full overflow-y-auto bg-white px-6 py-6 sm:max-w-sm sm:ring-1 sm:ring-gray-900/10">
+                    <div className="flex items-center justify-between">
+                        <Link to="/" className="-m-1.5 p-1.5">
+                            <span className="sr-only">Your Company</span>
+                            <img alt="" src="https://tailwindui.com/img/logos/mark.svg?color=indigo&shade=600" className="h-8 w-auto" />
+                        </Link>
+                        <button
+                            type="button"
+                            onClick={() => setMobileMenuOpen(false)}
+                            className="-m-2.5 rounded-md p-2.5 text-gray-700"
+                        >
+                            <span className="sr-only">Close menu</span>
+                            <XMarkIcon aria-hidden="true" className="h-6 w-6" />
+                        </button>
+                    </div>
+                    <div className="mt-6 flow-root">
+                        <div className="-my-6 divide-y divide-gray-500/10">
+                            <div className="space-y-2 py-6">
+                                <Disclosure as="div" className="-mx-3">
+                                    <DisclosureButton className="group flex w-full items-center justify-between rounded-lg py-2 pl-3 pr-3.5 text-base font-semibold leading-7 text-gray-900 hover:bg-gray-50">
+                                        Product
+                                        <ChevronDownIcon aria-hidden="true" className="h-5 w-5 flex-none group-data-[open]:rotate-180" />
+                                    </DisclosureButton>
+                                    <DisclosurePanel className="mt-2 space-y-2">
+                                        {[...products, ...callsToAction].map((item) => (
+                                            <DisclosureButton
+                                                key={item.name}
+                                                as="a"
+                                                href={item.href}
+                                                className="block rounded-lg py-2 pl-6 pr-3 text-sm font-semibold leading-7 text-gray-900 hover:bg-gray-50"
+                                            >
+                                                {item.name}
+                                            </DisclosureButton>
+                                        ))}
+                                    </DisclosurePanel>
+                                </Disclosure>
+                                <Link to="/features" className="-mx-3 block rounded-lg px-3 py-2 text-base font-semibold leading-7 text-gray-900 hover:bg-gray-50">
+                                    Features
+                                </Link>
+                                <Link to="/marketplace" className="-mx-3 block rounded-lg px-3 py-2 text-base font-semibold leading-7 text-gray-900 hover:bg-gray-50">
+                                    Marketplace
+                                </Link>
+                                <Link to="/company" className="-mx-3 block rounded-lg px-3 py-2 text-base font-semibold leading-7 text-gray-900 hover:bg-gray-50">
+                                    Company
+                                </Link>
+                            </div>
+                            <div className="py-6">
+                                <Link to="/login" className="-mx-3 block rounded-lg px-3 py-2.5 text-base font-semibold leading-7 text-gray-900 hover:bg-gray-50">
+                                    Log in
+                                </Link>
+                            </div>
+                        </div>
+                    </div>
+                </DialogPanel>
+            </Dialog>
+        </header>
+    );
+}
